+++ src/logic/configure.test.ts
@@ -3,6 +3,7 @@ import {
   ApiKeyPlacement,
   HttpScheme,
   ok,
+  OnFail,
   ProviderJson,
   SecurityType,
   SuperJson,
@@ -278,7 +279,14 @@ describe('Configure CLI logic', () => {
         .mockResolvedValue(undefined);
 
       await expect(
-        installProvider('some/path', providerName, mockProfileId)
+        installProvider({
+          superPath: 'some/path',
+          provider: providerName,
+          profileId: mockProfileId,
+          defaults: {
+            testUseCase: { retryPolicy: OnFail.CIRCUIT_BREAKER },
+          },
+        })
       ).resolves.toBeUndefined();
 
       expect(loadSpy).toHaveBeenCalledTimes(1);
@@ -290,7 +298,7 @@ describe('Configure CLI logic', () => {
       expect(addProfileProviderSpy).toHaveBeenCalledWith(
         mockProfileId,
         providerName,
-        {}
+        { testUseCase: { retryPolicy: OnFail.CIRCUIT_BREAKER } }
       );
 
       expect(addProviderSpy).toHaveBeenCalledTimes(1);
@@ -341,8 +349,14 @@ describe('Configure CLI logic', () => {
       mocked(readFile).mockResolvedValue(JSON.stringify(mockProviderJson));
 
       await expect(
-        installProvider('some/path', providerName, mockProfileId, {
-          local: true,
+        installProvider({
+          superPath: 'some/path',
+          provider: providerName,
+          profileId: mockProfileId,
+          defaults: undefined,
+          options: {
+            local: true,
+          },
         })
       ).resolves.toBeUndefined();
 
@@ -395,8 +409,14 @@ describe('Configure CLI logic', () => {
       mocked(readFile).mockRejectedValue(new Error('test'));
 
       await expect(
-        installProvider('some/path', providerName, mockProfileId, {
-          local: true,
+        installProvider({
+          superPath: 'some/path',
+          provider: providerName,
+          profileId: mockProfileId,
+          defaults: undefined,
+          options: {
+            local: true,
+          },
         })
       ).rejects.toEqual(new CLIError('test'));
 
@@ -419,7 +439,11 @@ describe('Configure CLI logic', () => {
         .mockResolvedValue(ok(mockSuperJson));
 
       await expect(
-        installProvider('some/path', providerName, mockProfileId)
+        installProvider({
+          superPath: 'some/path',
+          provider: providerName,
+          profileId: mockProfileId,
+        })
       ).rejects.toEqual(
         new CLIError(
           `❌ profile ${mockProfileId} not found in some/path. Forgot to install?`
@@ -465,7 +489,11 @@ describe('Configure CLI logic', () => {
         .mockResolvedValue(undefined);
 
       await expect(
-        installProvider('some/path', providerName, mockProfileId)
+        installProvider({
+          superPath: 'some/path',
+          provider: providerName,
+          profileId: mockProfileId,
+        })
       ).resolves.toBeUndefined();
 
       expect(loadSpy).toHaveBeenCalledTimes(1);