+++ CHANGELOG.md
@@ -6,8 +6,19 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Multiple usecases in interactive install command
+- Repeated use of interactive install command
+- Set provider failover in interactive install command
+- Set provider retry policies in interactive install command
+- Providers priority array
+
+### Fixed
+- Package manager path resolution
 
 ## [0.0.19] - 2021-07-08
+### Fixed
+- Last provider selection in interactive install command
 
 ## [0.0.18] - 2021-06-15
 ### Added